+++ src/features/ControlPanel/SkipInterval.tsx
